+++ client/components/Main/index.js
@@ -1,8 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+import './style.css'
+
 const Main = ({ children }) => (
-  <div>
+  <div id="content-wrapper">
     {children}
   </div>
 )
