+++ client/components/Main/style.css
@@ -1,0 +1,5 @@
+#content-wrapper {
+  display: flex;
+  flex-direction: column;
+  margin: 0 3em;
+}
