+++ client/pages/index.js
@@ -1,0 +1,2 @@
+export { default as RandomLargeSum } from './RandomLargeSum.js'
+export { default as RandomLargeSumContributor } from './RandomLargeSumContributor.js'
