+++ client/style.css
@@ -1,10 +1,6 @@
   #app-wrapper {}
 
-  #content-wrapper {
-    display: flex;
-    flex-direction: column;
-    margin: 0 3em;
-  }
+
 
   .algo-name-header-wrapper {
     margin-bottom: 1em;
